wingui: Fix painting of edges at startup; improve program icon loading
- Previously, if a WinGUI window was created maximized, the right and
bottom edges of the window (not covered by the curses screen) would not
be filled in. This patch fixes this.
- Changed the icon loading function to use EnumResourceNames() and a
corresponding callback function with LoadIcon() to load the first icon
in the application's executable. ExtractIcon() does not properly load
icon groups, resulting in a wrong icon size being used in the window's
title bar. This fix also removes any dependency on shell32.dll or
shellapi.h.
- Replaced an instance of GetModuleHandleA() with the GetModuleHandle()
macro to avoid unnecessary importing of both A and W versions in a
Unicode build.