wingui: Move window procedure to separate thread to fix resize/move
This fixes issue #156, wherein one's application was hung during a
window move or resize, preventing it from processing KEY_RESIZE events
during a resize. This is because the window procedure was executing in
the same thread as the application, and execution outside the window
procedure in its thread is blocked while it is in resize/move mode.