wingui: Use WS_EX_CLIENTEDGE extended style with window
This provides a cleaner appearance with the Windows "classic" style by
framing the client area with a recessed border, giving the window
frame a more proper "bezel"/"frame" appearance.

Also fixed a bug where disabling the menu bar would set the window's
menu bar to a new empty one instead of the proper method of setting it
to NULL, and changed the variable types used for the window styles
from int to the proper DWORD type.