wincon & wingui: Play beeps synchronously so multiple beeps work
Previously, PDC_beep() in WinCon and WinGUI was implemented with a
call to MessageBeep(0xFFFFFFFF), which plays the Windows default beep
sound (or a standard beep on the PC speaker as a fallback). This
function plays the sound asynchronously, resulting in multiple beeps
in a row (such as with the "TwoBeeps" option in TUIDEMO) overlapping
instead of producing the intended results.

This patch changes it to use PlaySound() to play the Windows default
beep synchronously. If PlaySound() fails, it will call Beep(800, 200)
as a fallback to play a standard PC speaker beep. The makefiles for
Digital Mars, MinGW/GCC, and Visual C++ were also updated to link the
required winmm library.