[FIX]: key modifiers were cleared without a key event.
Key modifiers were incorrectly cleared when checking for an event. Now they are updated only after a key event.

Before this commit, for instance, the num lock status, along with the rest of the modifiers, was cleared when polling for a new key event. After this commit, the modifiers are not cleared without a correct update of the status, only when a new key is pressed. This aligns with the key behavior in other flavors such as wingui.