SDL1 and x11new now set the 'repeat' modifier bit correctly.
In SDL1,  if we get a key down event,  a subsequent key down event for
the same key with no intervening key up event is interpreted as meaning
a repeat,  and the flag is set.

In X11 (new),  if we get a key down message right after a key up message for
the same key,  with identical time stamps,  the key up is a repeat.