getch() triggered an assert() if a multi-byte character was entered. Fixes issue #335,  reported by Robin Haberkorn.
Previous versions of PDC_wcstombs(),  when compiled with PDC_FORCE_UTF8,
did minimal error-checking for buffer overflow.  Fixing this revealed
that getch() mis-used the function,  assuming that the last argument
was a count of wide characters input (of which there would be one),
rather than an output buffer size (which would be eight bytes).