If an application #defines PDC_NCMOUSE,  the behavior of mouse function is fully compatible with ncurses.  Wheel up/down is mapped to buttons 4/5 pressed.
If PC_NCMOUSE is #defined by a program,  that fact is conveyed to
the library when nc_maskmouse() is first called.  This signals to
the PDCursesMod library that vertical mouse wheel events should (a
la ncurses) be translated to buttons 4 and 5,  and tilt-wheel mouse
events should be ignored.  Fix for issue #255.