Extracted logic for software blinking and PDC_gotoyx() into common/blink.c
All ports except DOS,  console OS/2,  and VT have nearly similar
functions to blink text and to redraw the cursor,  and to handle
the PDC_gotoyx() function.  The relevant logic from the framebuffer
port is now in common/blink.c,  and will soon be used on most or
all of the other non-hardware-blinking platforms.